requirements.txt
chromadb
zhipuai
pdfplumber
streamlit
python-dotenv
python-docx